+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add support for date filters and `finished_at` ordering in `Deployments::all`
 * Add support for listing merge requests associated with a deployment
 * Add support for `with_custom_attributes` and `with_projects` in `Groups::show`
+* Add support for project remote mirror read endpoints
 
 
 ## [12.0.0] - 2025-02-23